Macro code organization cleanup (#115)
* Improve error handling

* Group macro code into file modules
